refactor(AssetUploader): extract 'imageable' methods to concern
